+++ src/main/java/kitchenpos/application/MenuService.java
@@ -3,23 +3,23 @@ package kitchenpos.application;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
-import kitchenpos.domain.Quantity;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.menuproduct.MenuProduct;
+import kitchenpos.domain.product.Product;
+import kitchenpos.domain.quantity.Quantity;
 import kitchenpos.domain.price.Price;
 import kitchenpos.domain.productquantity.ProductQuantities;
 import kitchenpos.domain.productquantity.ProductQuantity;
-import kitchenpos.dto.menu.MenuProductRequest;
-import kitchenpos.dto.menu.MenuProductResponse;
+import kitchenpos.dto.menuproduct.MenuProductRequest;
+import kitchenpos.dto.menuproduct.MenuProductResponse;
 import kitchenpos.dto.menu.MenuRequest;
 import kitchenpos.dto.menu.MenuResponse;
 import kitchenpos.exception.NotFoundException;
-import kitchenpos.repository.MenuGroupRepository;
-import kitchenpos.repository.MenuProductRepository;
-import kitchenpos.repository.MenuRepository;
-import kitchenpos.repository.ProductRepository;
+import kitchenpos.domain.menugroup.MenuGroupRepository;
+import kitchenpos.domain.menuproduct.MenuProductRepository;
+import kitchenpos.domain.menu.MenuRepository;
+import kitchenpos.domain.product.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
