+++ src/main/java/kitchenpos/application/OrderService.java
@@ -3,22 +3,22 @@ package kitchenpos.application;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.Order;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.order.Order;
 import kitchenpos.domain.orderlineitem.OrderLineItem;
 import kitchenpos.domain.orderlineitem.OrderLineItems;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.Quantity;
+import kitchenpos.domain.quantity.Quantity;
 import kitchenpos.dto.order.OrderRequest;
 import kitchenpos.dto.order.OrderResponse;
 import kitchenpos.dto.orderlineitem.OrderLineItemRequest;
 import kitchenpos.dto.orderlineitem.OrderLineItemResponse;
 import kitchenpos.exception.NotFoundException;
-import kitchenpos.repository.MenuRepository;
-import kitchenpos.repository.OrderLineItemRepository;
-import kitchenpos.repository.OrderRepository;
-import kitchenpos.repository.OrderTableRepository;
+import kitchenpos.domain.menu.MenuRepository;
+import kitchenpos.domain.orderlineitem.OrderLineItemRepository;
+import kitchenpos.domain.order.OrderRepository;
+import kitchenpos.domain.ordertable.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
