+++ src/main/java/kitchenpos/domain/menu/MenuRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.repository;
+package kitchenpos.domain.menu;
 
-import kitchenpos.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
