+++ src/main/java/kitchenpos/domain/menuproduct/MenuProductRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.domain.menuproduct;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuProductRepository extends JpaRepository<MenuProduct, Long> {
