+++ src/main/java/kitchenpos/domain/orderlineitem/OrderLineItemRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.repository;
+package kitchenpos.domain.orderlineitem;
 
 import java.util.List;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.orderlineitem.OrderLineItem;
+import kitchenpos.domain.order.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
