+++ src/main/java/kitchenpos/domain/product/Product.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.product;
 
 import java.util.Objects;
 import javax.persistence.Column;
