+++ src/main/java/kitchenpos/domain/quantity/Quantity.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.quantity;
 
 import java.util.Objects;
 import javax.persistence.Column;
