+++ src/main/java/kitchenpos/domain/tablegroup/TableGroupRepository.java
@@ -1,6 +1,5 @@
-package kitchenpos.repository;
+package kitchenpos.domain.tablegroup;
 
-import kitchenpos.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
