+++ src/main/java/kitchenpos/dto/menu/MenuResponse.java
@@ -2,7 +2,8 @@ package kitchenpos.dto.menu;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.Menu;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.dto.menuproduct.MenuProductResponse;
 
 public class MenuResponse {
