+++ src/main/java/kitchenpos/dto/order/OrderResponse.java
@@ -2,8 +2,8 @@ package kitchenpos.dto.order;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.dto.orderlineitem.OrderLineItemResponse;
 
 public class OrderResponse {
