+++ src/test/java/kitchenpos/domain/productquantity/ProductQuantitiesTest.java
@@ -3,8 +3,8 @@ package kitchenpos.domain.productquantity;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import kitchenpos.domain.Product;
-import kitchenpos.domain.Quantity;
+import kitchenpos.domain.product.Product;
+import kitchenpos.domain.quantity.Quantity;
 import kitchenpos.domain.price.Price;
 import kitchenpos.exception.InvalidArgumentException;
 import org.junit.jupiter.api.DisplayName;
