+++ src/test/java/kitchenpos/integration/MenuIntegrationTest.java
@@ -11,11 +11,11 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import kitchenpos.config.CustomParameterizedTest;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
-import kitchenpos.dto.menu.MenuProductRequest;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.menuproduct.MenuProduct;
+import kitchenpos.domain.product.Product;
+import kitchenpos.dto.menuproduct.MenuProductRequest;
 import kitchenpos.dto.menu.MenuRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
