+++ src/test/java/kitchenpos/integration/OrderIntegrationTest.java
@@ -11,17 +11,17 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import java.util.Collections;
 import java.util.List;
 import kitchenpos.config.CustomParameterizedTest;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.Order;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menugroup.MenuGroup;
+import kitchenpos.domain.order.Order;
 import kitchenpos.domain.orderlineitem.OrderLineItem;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.Quantity;
+import kitchenpos.domain.quantity.Quantity;
 import kitchenpos.dto.order.OrderRequest;
 import kitchenpos.dto.orderlineitem.OrderLineItemRequest;
 import kitchenpos.exception.NotFoundException;
-import kitchenpos.repository.OrderLineItemRepository;
+import kitchenpos.domain.orderlineitem.OrderLineItemRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
