+++ src/test/java/kitchenpos/integration/TableGroupIntegrationTest.java
@@ -12,8 +12,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import kitchenpos.config.CustomParameterizedTest;
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.OrderStatus;
+import kitchenpos.domain.tablegroup.TableGroup;
 import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.dto.ordertable.OrderTableRequest;
 import kitchenpos.dto.tablegroup.TableGroupRequest;
